store existing stats in redis, more refactoring
