+++ pkg/utils/stats/stats.go
@@ -2,206 +2,24 @@ package stats
 
 import (
 	"io"
-	"sync/atomic"
 	"time"
 
+	livekit "github.com/livekit/livekit-server/proto"
 	"github.com/pion/interceptor"
 	"github.com/pion/rtcp"
 	"github.com/pion/rtp"
 	"github.com/pion/transport/packetio"
-	"github.com/prometheus/client_golang/prometheus"
 )
 
 const livekitNamespace = "livekit"
 
 var (
-	promLabels  = []string{"direction"}
-	packetTotal = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "packet",
-		Name:      "total",
-	}, promLabels)
-	packetBytes = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "packet",
-		Name:      "bytes",
-	}, promLabels)
-	nackTotal = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "nack",
-		Name:      "total",
-	}, promLabels)
-	pliTotal = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "pli",
-		Name:      "total",
-	}, promLabels)
-	firTotal = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "fir",
-		Name:      "total",
-	}, promLabels)
-	roomTotal = prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "room",
-		Name:      "total",
-	})
-	roomDuration = prometheus.NewHistogram(prometheus.HistogramOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "room",
-		Name:      "duration_seconds",
-		Buckets: []float64{
-			5, 10, 60, 5 * 60, 10 * 60, 30 * 60, 60 * 60, 2 * 60 * 60, 5 * 60 * 60, 10 * 60 * 60,
-		},
-	})
-	participantTotal = prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "participant",
-		Name:      "total",
-	})
-	trackPublishedTotal = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "track",
-		Name:      "published_total",
-	}, []string{"kind"})
-	trackSubscribedTotal = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Namespace: livekitNamespace,
-		Subsystem: "track",
-		Name:      "subscribed_total",
-	}, []string{"kind"})
+	promLabels = []string{"direction"}
 )
 
 func init() {
-	prometheus.MustRegister(packetTotal)
-	prometheus.MustRegister(packetBytes)
-	prometheus.MustRegister(nackTotal)
-	prometheus.MustRegister(pliTotal)
-	prometheus.MustRegister(firTotal)
-	prometheus.MustRegister(roomTotal)
-	prometheus.MustRegister(roomDuration)
-	prometheus.MustRegister(participantTotal)
-	prometheus.MustRegister(trackPublishedTotal)
-	prometheus.MustRegister(trackSubscribedTotal)
-}
-
-// RoomStatsReporter is created for each room
-type RoomStatsReporter struct {
-	roomName  string
-	startedAt time.Time
-	Incoming  *PacketStats
-	Outgoing  *PacketStats
-}
-
-func NewRoomStatsReporter(roomName string) *RoomStatsReporter {
-	return &RoomStatsReporter{
-		roomName: roomName,
-		Incoming: newPacketStats(roomName, "incoming"),
-		Outgoing: newPacketStats(roomName, "outgoing"),
-	}
-}
-
-func (r *RoomStatsReporter) RoomStarted() {
-	r.startedAt = time.Now()
-	roomTotal.Add(1)
-}
-
-func (r *RoomStatsReporter) RoomEnded() {
-	if !r.startedAt.IsZero() {
-		roomDuration.Observe(float64(time.Now().Sub(r.startedAt)) / float64(time.Second))
-	}
-	roomTotal.Sub(1)
-}
-
-func (r *RoomStatsReporter) AddParticipant() {
-	participantTotal.Add(1)
-}
-
-func (r *RoomStatsReporter) SubParticipant() {
-	participantTotal.Sub(1)
-}
-
-func (r *RoomStatsReporter) AddPublishedTrack(kind string) {
-	trackPublishedTotal.WithLabelValues(kind).Add(1)
-}
-
-func (r *RoomStatsReporter) SubPublishedTrack(kind string) {
-	trackPublishedTotal.WithLabelValues(kind).Sub(1)
-}
-
-func (r *RoomStatsReporter) AddSubscribedTrack(kind string) {
-	trackSubscribedTotal.WithLabelValues(kind).Add(1)
-}
-
-func (r *RoomStatsReporter) SubSubscribedTrack(kind string) {
-	trackSubscribedTotal.WithLabelValues(kind).Sub(1)
-}
-
-type PacketStats struct {
-	roomName  string
-	direction string // incoming or outgoing
-
-	PacketBytes uint64 `json:"packetBytes"`
-	PacketTotal uint64 `json:"packetTotal"`
-	NackTotal   uint64 `json:"nackTotal"`
-	PLITotal    uint64 `json:"pliTotal"`
-	FIRTotal    uint64 `json:"firTotal"`
-}
-
-func newPacketStats(room, direction string) *PacketStats {
-	return &PacketStats{
-		roomName:  room,
-		direction: direction,
-	}
-}
-
-func (s *PacketStats) IncrementBytes(bytes uint64) {
-	packetBytes.WithLabelValues(s.direction).Add(float64(bytes))
-	atomic.AddUint64(&s.PacketBytes, bytes)
-}
-
-func (s *PacketStats) IncrementPackets(count uint64) {
-	packetTotal.WithLabelValues(s.direction).Add(float64(count))
-	atomic.AddUint64(&s.PacketTotal, count)
-}
-
-func (s *PacketStats) IncrementNack(count uint64) {
-	nackTotal.WithLabelValues(s.direction).Add(float64(count))
-	atomic.AddUint64(&s.NackTotal, count)
-}
-
-func (s *PacketStats) IncrementPLI(count uint64) {
-	pliTotal.WithLabelValues(s.direction).Add(float64(count))
-	atomic.AddUint64(&s.PLITotal, count)
-}
-
-func (s *PacketStats) IncrementFIR(count uint64) {
-	firTotal.WithLabelValues(s.direction).Add(float64(count))
-	atomic.AddUint64(&s.FIRTotal, count)
-}
-
-func (s *PacketStats) HandleRTCP(pkts []rtcp.Packet) {
-	for _, rtcpPacket := range pkts {
-		switch rtcpPacket.(type) {
-		case *rtcp.TransportLayerNack:
-			s.IncrementNack(1)
-		case *rtcp.PictureLossIndication:
-			s.IncrementPLI(1)
-		case *rtcp.FullIntraRequest:
-			s.IncrementFIR(1)
-		}
-	}
-}
-
-func (s PacketStats) Copy() *PacketStats {
-	return &PacketStats{
-		roomName:    s.roomName,
-		direction:   s.direction,
-		PacketBytes: atomic.LoadUint64(&s.PacketBytes),
-		PacketTotal: atomic.LoadUint64(&s.PacketTotal),
-		NackTotal:   atomic.LoadUint64(&s.NackTotal),
-		PLITotal:    atomic.LoadUint64(&s.PLITotal),
-		FIRTotal:    atomic.LoadUint64(&s.FIRTotal),
-	}
+	initPacketStats()
+	initRoomStatsReporter()
 }
 
 // StatsBufferWrapper wraps a buffer factory so we could get information on
@@ -264,4 +82,13 @@ func (s *StatsInterceptor) BindLocalStream(_ *interceptor.StreamInfo, writer int
 		s.reporter.Outgoing.IncrementBytes(uint64(len(payload)))
 		return writer.Write(header, payload, attributes)
 	})
+}
+
+func UpdateCurrentNodeStats(nodeStats *livekit.NodeStats) {
+	nodeStats.NumClients = uint32(participantTotal)
+	nodeStats.NumRooms = uint32(roomTotal)
+	nodeStats.NumTracksIn = uint32(trackPublishedTotal)
+	nodeStats.NumTracksOut = uint32(trackSubscribedTotal)
+
+	nodeStats.UpdatedAt = time.Now().Unix()
 }
